refactor: Rename MakeEmptyMetricsData
Rename MakeEmptyMetricsData as in go default name for constructors
should be with schema like NewStructName and name of this function is
not in go way.

Reference to Effective go:

> Similarly, the function to make new instances of ring.Ring—which is
the definition of a constructor in Go—would normally be called NewRing,
but since Ring is the only type exported by the package, and since the
package is called ring, it's called just New, which clients of the
package see as ring.New. Use the package structure to help you choose
good names.